[refact] Move BusMmio type into caliptra-emu-bus.
